(general) fixing emitted warnings

+++ src/Mod/RuleBaseFacility.h
@@ -56,7 +56,7 @@ private:
 	int _mindPower;
 	int _size, _buildCost, _refundValue, _buildTime, _monthlyCost;
 	std::map<std::string, std::pair<int, int> > _buildCostItems;
-	int _storage, _personnel, _aliens, _crafts, _labs, _workshops, _psiLabs;
+	size_t _storage, _personnel, _aliens, _crafts, _labs, _workshops, _psiLabs;
 	bool _spriteEnabled;
 	int _sightRange, _sightChance;
 	int _radarRange, _radarChance, _defense, _hitRatio, _fireSound, _hitSound, _placeSound;
